backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymupdf==1.23.6
anthropic==0.8.0
pinecone-client==2.2.4
python-multipart==0.0.6
python-dotenv==1.0.0
sentence-transformers==2.3.1
numpy==1.26.2
pillow==10.1.0